requirements.txt
flask
openpyxl
pillow
pytesseract